pyproject.toml
[tool.poetry]
name = "ai-agent-educatif"
version = "0.1.0"
description = "Agent IA éducatif avec démonstration POO et interface Streamlit"
authors = ["Étudiant <etudiant@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.28.0"
plotly = "^5.17.0"
openai = "^1.0.0"
python-dotenv = "^1.0.0"
fastapi = "^0.104.0"
uvicorn = "^0.24.0"
httpx = "^0.25.0"

[tool.poetry.group.dev.dependencies]
black = "^23.0.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"